+++ lib/hako/schedulers/ecs.rb
@@ -60,9 +60,21 @@ module Hako
         end
         @oneshot_notification_prefix = options.fetch('oneshot_notification_prefix', nil)
         if options.key?('deployment_configuration')
+          deployment_configuration = options.fetch('deployment_configuration')
+
           @deployment_configuration = {}
+          if deployment_configuration.key?('deployment_circuit_breaker')
+            circuit_breaker = deployment_configuration.fetch('deployment_circuit_breaker')
+            if circuit_breaker.fetch('rollback', nil)
+              validation_error!('rollback option of deployment_circuit_breaker is not supported because Hako handles rollback instead of the circuit breaker')
+            end
+            @deployment_configuration[:deployment_circuit_breaker] = {
+              enable: circuit_breaker.fetch('enable'),
+              rollback: false,
+            }
+          end
           %i[maximum_percent minimum_healthy_percent].each do |key|
-            @deployment_configuration[key] = options.fetch('deployment_configuration')[key.to_s]
+            @deployment_configuration[key] = deployment_configuration[key.to_s]
           end
         else
           @deployment_configuration = nil
@@ -117,6 +129,7 @@ module Hako
           @service_discovery = EcsServiceDiscovery.new(options.fetch('service_discovery'), @region, dry_run: @dry_run)
         end
         @tags = options.fetch('tags', {}).map { |k, v| { key: k, value: v.to_s } }
+        @enable_execute_command = options.fetch('enable_execute_command', false)
 
         @started_at = nil
         @container_instance_arn = nil
@@ -724,6 +737,7 @@ module Hako
           system_controls: container.system_controls,
           repository_credentials: container.repository_credentials,
           resource_requirements: container.resource_requirements,
+          firelens_configuration: container.firelens_configuration,
         }
       end
 
@@ -745,6 +759,7 @@ module Hako
           platform_version: @platform_version,
           network_configuration: @network_configuration,
           propagate_tags: 'TASK_DEFINITION',
+          enable_execute_command: @enable_execute_command,
         )
         result.failures.each do |failure|
           Hako.logger.error("#{failure.arn} #{failure.reason}")
@@ -941,6 +956,7 @@ module Hako
           platform_version: @platform_version,
           network_configuration: @network_configuration,
           health_check_grace_period_seconds: @health_check_grace_period_seconds,
+          enable_execute_command: @enable_execute_command,
         }
         if @autoscaling
           # Keep current desired_count if autoscaling is enabled
@@ -994,6 +1010,7 @@ module Hako
           network_configuration: @network_configuration,
           health_check_grace_period_seconds: @health_check_grace_period_seconds,
           propagate_tags: 'TASK_DEFINITION',
+          enable_execute_command: @enable_execute_command,
         }
         if @scheduling_strategy != 'DAEMON'
           params[:desired_count] = 0
@@ -1056,13 +1073,14 @@ module Hako
           Hako.logger.debug "  latest_event_id=#{latest_event_id}, deployments=#{s.deployments}"
           no_active = s.deployments.all? { |d| d.status != 'ACTIVE' }
           primary = s.deployments.find { |d| d.status == 'PRIMARY' }
-          if primary.desired_count * 2 < @started_task_ids.size
+          if primary.rollout_state == 'FAILED'
+            Hako.logger.error("New deployment is failing: #{primary.rollout_state_reason}")
+            report_task_diagnostics(service.cluster_arn, @started_task_ids)
+            return false
+          end
+          if primary.desired_count * 2 < @started_task_ids.size && !ecs_circuit_breaker_enabled?
             Hako.logger.error('Some started tasks are stopped. It seems new deployment is failing to start')
-            @started_task_ids.each_slice(100) do |task_ids|
-              ecs_client.describe_tasks(cluster: service.cluster_arn, tasks: task_ids).tasks.each do |task|
-                report_task_diagnostics(task)
-              end
-            end
+            report_task_diagnostics(service.cluster_arn, @started_task_ids)
             return false
           end
           primary_ready = primary && primary.running_count == primary.desired_count
@@ -1072,6 +1090,11 @@ module Hako
             sleep 1
           end
         end
+      end
+
+      # @return [Boolean]
+      def ecs_circuit_breaker_enabled?
+        @deployment_configuration&.dig(:deployment_circuit_breaker, :enable)
       end
 
       # @param [Array<Aws::ECS::Types::ServiceEvent>] events
@@ -1091,13 +1114,17 @@ module Hako
         message.slice(TASK_ID_RE, 1)
       end
 
-      # @param [Aws::ECS::Types::Task] task
+      # @param [Array<String>] task_ids
       # @return [nil]
-      def report_task_diagnostics(task)
-        Hako.logger.error("task_definition_arn=#{task.task_definition_arn} last_status=#{task.last_status}")
-        Hako.logger.error("  stopped_reason: #{task.stopped_reason}")
-        task.containers.sort_by(&:name).each do |container|
-          Hako.logger.error("    Container #{container.name}: last_status=#{container.last_status} exit_code=#{container.exit_code.inspect} reason=#{container.reason.inspect}")
+      def report_task_diagnostics(cluster, task_ids)
+        task_ids.each_slice(100) do |batch|
+          ecs_client.describe_tasks(cluster: cluster, tasks: batch).tasks.each do |task|
+            Hako.logger.error("task_definition_arn=#{task.task_definition_arn} last_status=#{task.last_status}")
+            Hako.logger.error("  stopped_reason: #{task.stopped_reason}")
+            task.containers.sort_by(&:name).each do |container|
+              Hako.logger.error("    Container #{container.name}: last_status=#{container.last_status} exit_code=#{container.exit_code.inspect} reason=#{container.reason.inspect}")
+            end
+          end
         end
       end