+++ spec/hako/definition_loader_spec.rb
@@ -47,6 +47,23 @@ RSpec.describe Hako::DefinitionLoader do
       end
     end
 
+    context 'with depends_on' do
+      let(:fixture_name) { 'default_with_depends_on.jsonnet' }
+
+      it 'loads all containers' do
+        containers = definition_loader.load
+        expect(containers.keys).to match_array(%w[app redis memcached fluentd init init2])
+        expect(containers.values).to all(be_a(Hako::Container))
+      end
+
+      context 'with `with`' do
+        it 'loads specified definition and linked containers and dependent containers' do
+          containers = definition_loader.load(with: [])
+          expect(containers.keys).to match_array(%w[app redis memcached init init2])
+        end
+      end
+    end
+
     context 'with volumes_from' do
       let(:fixture_name) { 'default_with_volumes_from.jsonnet' }